Removed UnwindSafe bounds
After thinking about this more and more, I've come to realize that
forcing UnwindSafe is not the intention of the bound on catch_unwind.
This should have been evident to me by the fact that thread::spawn does
not require UnwindSafe, yet it catches panics. The key qualifier I wasn't
noticing was that the design of the trait is to prevent *easily*
observing invariant states.

Since this panic catch results in dropping and then subsequently closing
everything that was passed to it, it fits the same general shape as
thread::spawn, so I'm removing the bounds.